Atividade 5, comentários finais

+++ Atividade 5/atv5_novo/multi-app/client/src/components/TodoApp.jsx
@@ -5,7 +5,8 @@ import { useState, useEffect } from 'react';
 import axios from 'axios';
 // Importa os componentes estilizados para renderizar
 import { Container, Title, Input, Button, TaskList, TaskItem, EditInput } from '../assets/styles/TodoApp-styling'
-
+// Importa o hook para verificação do JWT
+import useVerifyJWT from "../assets/hooks/useVerifyJWT" // Hook para fazer validação do JWT
 // Define a URL da API que será usada para obter, adicionar, editar e excluir tarefas.
 const API_URL = 'http://localhost:3000/tasks';
 
@@ -18,6 +19,7 @@ const TodoApp = () => {
   const [editingTaskId, setEditingTaskId] = useState(null); // Estado para o id da tarefa que está sendo editada.
   const [editingTaskText, setEditingTaskText] = useState(''); // Estado para o texto da tarefa que está sendo editada.
 
+  const verifyJWT = useVerifyJWT() // Recebe a função updateToken do hook
   // Usa o hook useEffect para buscar as tarefas quando o componente é montado.
   useEffect(() => {
     fetchTasks();
@@ -26,8 +28,6 @@ const TodoApp = () => {
   // Função que busca as tarefas da API e atualiza o estado com as tarefas recebidas.
   const fetchTasks = async () => {
     const response = await axios.get(API_URL); // Faz uma requisição GET para obter as tarefas.
-    
-    
     if (JSON.stringify(response.data) != JSON.stringify(tasks)) {
       setTasks(response.data); // Atualiza o estado com os dados recebidos.
     }
@@ -36,37 +36,44 @@ const TodoApp = () => {
   // Função que adiciona uma nova tarefa.
   const addTask = async () => {
     try {
-      if (task) { 
-        const newTask = { text: task };
-        
-        // Espera a resposta da adição da nova tarefa.
-        const response = await axios.post(API_URL, newTask);
-        
-        // Atualiza o estado localmente com a nova tarefa.
-        setTasks(prevTasks => [...prevTasks, response.data]);
-  
-        // Limpa o campo de entrada.
-        setTask(''); 
+      if (task) {
+        const tokenStatus = await verifyJWT() // Verifica o token
+        if (tokenStatus == true) {
+          const newTask = { text: task };
 
+          // Espera a resposta da adição da nova tarefa.
+          const response = await axios.post(API_URL, newTask);
+
+          // Atualiza o estado localmente com a nova tarefa.
+          setTasks(prevTasks => [...prevTasks, response.data]);
+
+          // Limpa o campo de entrada.
+          setTask('');
+        } else {
+          location.reload()
+        }
       }
-    } catch (error) { 
+    } catch (error) {
       console.error('Erro ao adicionar tarefa:', error);
-    } finally {
-      console.log('Tarefa adicionada com sucesso.');
     }
   };
-  
-  
+
+
   // Função que exclui uma tarefa.
   const deleteTask = async (id) => {
     try {
+      const tokenStatus = await verifyJWT()
+      if (tokenStatus == true) {
         await axios.delete(`${API_URL}/${id}`); // Faz uma requisição DELETE para excluir a tarefa com o id fornecido.
         setTasks(prevTasks => prevTasks.filter(task => task.id !== id)); // Atualiza o estado removendo a tarefa excluída.
+      } else {
+        location.reload()
+      }
     } catch (error) {
-        console.error('Failed to delete task:', error);
-        // Opcional: Adicionar tratamento de erro, como mostrar uma notificação ao usuário
+      console.error('Failed to delete task:', error);
+      // Opcional: Adicionar tratamento de erro, como mostrar uma notificação ao usuário
     }
-};
+  };
 
   // Função que inicia o processo de edição de uma tarefa.
   const editTask = (id, text) => {
@@ -76,13 +83,21 @@ const TodoApp = () => {
 
   // Função que atualiza uma tarefa existente.
   const updateTask = async (id) => {
-    const updatedTask = { text: editingTaskText }; // Cria um objeto de tarefa com o texto atualizado.
+    try {
+      const tokenStatus = await verifyJWT()
+      if (tokenStatus == true) {
+        const updatedTask = { text: editingTaskText }; // Cria um objeto de tarefa com o texto atualizado.
     /*const response = */await axios.put(`${API_URL}/${id}`, updatedTask); // Faz uma requisição PUT para atualizar a tarefa.
-    setTasks(tasks.map(task => (task.id === id ? { ...task, text: editingTaskText } : task))); // Atualiza o estado com a tarefa modificada.
-    // setTasks(response); // Atualiza o estado com a tarefa modificada.
-    setEditingTaskId(null); // Limpa o id da tarefa em edição.
-    setEditingTaskText(''); // Limpa o texto da tarefa em edição.
-    console.log(tasks)
+        setTasks(tasks.map(task => (task.id === id ? { ...task, text: editingTaskText } : task))); // Atualiza o estado com a tarefa modificada.
+        // setTasks(response); // Atualiza o estado com a tarefa modificada.
+        setEditingTaskId(null); // Limpa o id da tarefa em edição.
+        setEditingTaskText(''); // Limpa o texto da tarefa em edição.
+      } else {
+        location.reload()
+      }
+    } catch (error) {
+      console.error(error)
+    }
   };
 
   // Retorna o JSX que define o layout e comportamento do componente.
@@ -95,7 +110,7 @@ const TodoApp = () => {
         onChange={(e) => setTask(e.target.value)}
         placeholder="Add a new task"
       />
-      <Button onClick={() => {addTask()/*console.log(tasks)*/}}>Add Task</Button>
+      <Button onClick={() => { addTask() }}>Add Task</Button>
       <TaskList>
         {tasks.map((task) => (
           <TaskItem key={task.id}>
